Removed the outdated computerprogramming1, computerprogramming2 and computerprogramming3 source code directories. Updated all needed documents for the overall changes. To resolve issue #8
