+++ Java - Computer Programming 1/Java-Computer-Programming-1.md
@@ -1,0 +1,24 @@
+<!-- <p align="center"><img src="/md_assets/octocat.gif" alt="Logo" width="130" height="130"></p> -->
+<h3 align="center">Codes for Lab Lecture Exercises Used</h3>
+<p align="center"><em>Good 'old library/source codes that I've coded and used for lecturing CS/IT subjects and topics</em></p>
+<p align="center"><strong>You like the Repo? Don't forget to 🌟, 👁️, 🔱 and ❤️!</strong></p>
+<p align="center">
+   <img src="https://img.shields.io/badge/Purpose-EDUCATION/LEARNING-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Purpose-EDUCATION/LEARNING">
+   <img src="https://img.shields.io/badge/Version-3.0.0-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Version 3.0.0">
+   <img src="https://img.shields.io/badge/Lang-Java-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Lang-Java">
+   <img src="https://img.shields.io/badge/Database-None-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Database-None">
+   <img src="https://img.shields.io/badge/Contribute-CLOSED-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Contribute-CLOSED">
+   <img src="https://img.shields.io/badge/License-MIT-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="License MIT">
+   <img src="https://img.shields.io/badge/Last%20Update-12.03.2025-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Last Update 12.03.2025">
+   <a href="https://ko-fi.com/thenocturnaldevgypsy"><img src="https://img.shields.io/badge/Support%20me%20via%20Ko--Fi-%2300416a?logo=ko-fi&logoColor=white&color=%2300416a&textColor=white" alt="Support me via Ko-Fi"></a>
+<a href="https://github.com/sponsors/thenocturnaldevgypsy"><img src="https://custom-icon-badges.demolab.com/badge/Sponsor%20me%20at%20GitHub-%2300416a?logo=heart&logoColor=white&color=%2300416a&textColor=white" alt="Sponsor me at GitHub"></a>
+</p>
+
+## ![SECTION About this Directory](https://custom-icon-badges.demolab.com/badge/-About%20this%20Directory-2471AE?logo=repo&logoColor=white&labelColor=2471AE)
+
+`Java - Computer Programming 1` contains source codes I've coded and used for the said topic/subject during my BSCS/BSIT/BSCOE Assistant Professor Career.
+
+> [!NOTE]
+> *Still retrieving my old files for the codes, coming soon!*
+
+Back to [(main) README.md](../README.md)
