+++ Java - Computer Programming 2/src/ComputerProgramming2.java
@@ -1,0 +1,27 @@
+
+public class ComputerProgramming2 {
+  public static void main(String[] args) {
+
+    System.out.println("\n==============================================================================================================================");
+    System.out.println("+                                                                                                                            +");
+    System.out.println("+                                             Computer Programming 1 Lab Codes                                               +");
+    System.out.println("+                                                                                                                            +");
+    System.out.println("==============================================================================================================================");
+    System.out.println("+                                                                                                                            +");
+    System.out.println("+  Hello, I'm TheNocturnalDevGypsy, and looks like you've stumbled to one of my GitHub Repo: code-lab-lecture-exercises.     +");
+    System.out.println("+                                                                                                                            +");
+    System.out.println("+  It's a compilation of library/source codes that I've coded and used for lecturing CS/IT subjects and topics during my     +");
+    System.out.println("+  career as BSCS/BSIT/BSCOE Assistant Professor way back late 2007 to 2015. If you want to ask some questions and provide   +");
+    System.out.println("+  feedback, feel free to submit an issue at the repo with the label 'code-question', I'll try my best to answer.            +");
+    System.out.println("+                                                                                                                            +");
+    System.out.println("+  GitHub Repo: https://github.com/thenocturnaldevgypsy/code-lab-lecture-exercises                                           +");
+    System.out.println("+                                                                                                                            +");
+    System.out.println("+  If you find this helpful, please don't forget to STAR/FORK/WATCH the repo at GitHub!                                      +");
+    System.out.println("+                                                                                                                            +");
+    System.out.println("+  Or maybe buy me a coffee at Ko-Fi or sponsor me at GitHub? ;)                                                             +");
+    System.out.println("+                                                                                                                            +");
+    System.out.println("==============================================================================================================================\n\n");
+    System.out.println("This section is not implemented yet, still looking for the codes, Coming soon!\n\n");
+
+   }
+}
