+++ Java - Data Structure and Algorithm Analysis/Java-Data-Structure-and-Algorithm-Analysis.md
@@ -9,14 +9,14 @@
    <img src="https://img.shields.io/badge/Database-None-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Database-None">
    <img src="https://img.shields.io/badge/Contribute-CLOSED-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Contribute-CLOSED">
    <img src="https://img.shields.io/badge/License-MIT-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="License MIT">
-   <img src="https://img.shields.io/badge/Last%20Update-11.03.2025-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Last Update 11.03.2025">
+   <img src="https://img.shields.io/badge/Last%20Update-12.03.2025-%2300416a?logoColor=white&labelColor=%2300416a&color=%2324292e&textColor=white" alt="Last Update 12.03.2025">
    <a href="https://ko-fi.com/thenocturnaldevgypsy"><img src="https://img.shields.io/badge/Support%20me%20via%20Ko--Fi-%2300416a?logo=ko-fi&logoColor=white&color=%2300416a&textColor=white" alt="Support me via Ko-Fi"></a>
 <a href="https://github.com/sponsors/thenocturnaldevgypsy"><img src="https://custom-icon-badges.demolab.com/badge/Sponsor%20me%20at%20GitHub-%2300416a?logo=heart&logoColor=white&color=%2300416a&textColor=white" alt="Sponsor me at GitHub"></a>
 </p>
 
 ## ![SECTION About this Directory](https://custom-icon-badges.demolab.com/badge/-About%20this%20Directory-2471AE?logo=repo&logoColor=white&labelColor=2471AE)
 
-`Java - Data Structure and Algorithm Analysis` contains source codes I've coded and used for the said topic/subject during my CS/IT Professor Career.
+`Java - Data Structure and Algorithm Analysis` contains source codes I've coded and used for the said topic/subject during my BSCS/BSIT/BSCOE Assistant Professor Career.
 
 Back to [(main) README.md](../README.md)
